python-clip-service/requirements.txt
torch==2.0.1
torchvision==0.15.2
Flask==2.3.2
Pillow==10.0.0
requests==2.31.0
numpy==1.25.2 